+++ workflow/scripts/01.calculate_sumstats.py
@@ -49,15 +49,15 @@ for breaks_filename in snakemake.input.breakpoints:
     if "SFS" in breaks_filename:
         with warnings.catch_warnings():  # will prevent warning if file is empty
             warnings.simplefilter("ignore")
-            breaks_sfs = np.loadtxt(breaks_filename)
+            breaks_sfs = np.loadtxt(breaks_filename, dtype=float)
     elif "LD" in breaks_filename:
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            breaks_ld = np.loadtxt(breaks_filename)
+            breaks_ld = np.loadtxt(breaks_filename, dtype=float)
     elif "TM_WIN" in breaks_filename:
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
-            breaks_tm_win = np.loadtxt(breaks_filename)
+            breaks_tm_win = np.loadtxt(breaks_filename, dtype=float)
 
 # log
 with open(snakemake.log.log1, "a", encoding="utf-8") as logfile:
